Team player list Done

+++ baseball-ui/src/app/team/team-list/team-list.component.ts
@@ -13,8 +13,8 @@ import {
     bounceOutLeftAnimation,
     fadeInAnimation
 } from 'src/app/shared/animations/animation';
-import { MatTableDataSource, transformPanel } from '@angular/material';
-import { take } from 'rxjs/operators';
+import { MatTableDataSource, transformPanel, MatDialog } from '@angular/material';
+import { take, switchMap } from 'rxjs/operators';
 import {
     trigger,
     style,
@@ -29,6 +29,7 @@ import {
 } from '@angular/animations';
 import { AuthService } from 'src/app/modules/auth/auth.service';
 import { Roles } from 'src/app/shared/enums/roles.enum';
+import { DialogCustomComponent, DialogData } from 'src/app/shared/components/dialog-custom/dialog-custom.component';
 
 @Component({
     selector: 'app-team-list',
@@ -41,7 +42,7 @@ import { Roles } from 'src/app/shared/enums/roles.enum';
                     // The queries run in parallel with group function
                     query('.mat-header-cell', [
                         style({ transform: 'translateY(-20px' }),
-                        animate(20000)
+                        animate(1000)
                     ])
                     // { optional: true })
                 ])
@@ -74,34 +75,64 @@ import { Roles } from 'src/app/shared/enums/roles.enum';
     ]
 })
 export class TeamListComponent implements OnInit {
-    displayedColumns: string[] = ['firstName', 'lastName', 'age', 'phone', 'delete'];
+    displayedColumns: string[] = ['role', 'firstName', 'lastName', 'age', 'phone', 'delete'];
     public dataSource = new MatTableDataSource<User>();
+    userIsCoach: boolean;
 
     constructor(
         private teamService: TeamService,
-        public authService: AuthService) { }
+        public authService: AuthService,
+        private dialog: MatDialog
+    ) { }
 
     ngOnInit() {
         // Call get the teamMembers and subscribe to the
         // teamMembersObservable to update the list.
         this.teamService.getTeamMembers();
-        this.teamService.teamMembersObservable
-            .pipe(take(1))
-            .subscribe(teamMembersList => {
-                this.dataSource.data = teamMembersList;
-            });
+        // this.teamService.teamMembersObservable
+        //     .pipe(take(1))
+        //     .subscribe(teamMembersList => {
+        //         this.dataSource.data = teamMembersList;
+
+        //     });
 
         this.authService.user$.pipe(take(1)).subscribe((user) => {
-            // If not Coach do not show trash column 
+            // If current user is not a Coach do not show trash column
             if (!(user.roles.includes(Roles.COACH))) { this.displayedColumns.pop(); }
+
+            this.teamService.teamMembersObservable.pipe(take(1))
+                .subscribe(teamMembersList => {
+                    // If current user is coach remove the coach row from the list
+                    if ((user.roles.includes(Roles.COACH))) {
+                        const coachIndex = teamMembersList.findIndex(ele => {
+                            return (ele.roles.includes(Roles.COACH));
+                        });
+                        teamMembersList.splice(coachIndex, 1);
+                    }
+                    this.dataSource.data = teamMembersList;
+                });
         });
+
+        // this.authService.user$.pipe(switchMap((user) => {
+        //     if (!(user.roles.includes(Roles.COACH))) { this.displayedColumns.pop(); }
+        //     return user;
+        // }))
+
     }
 
     async onDelete(rowElement) {
-        await this.teamService.deleteUserFromTeam(rowElement.uid);
+        const dialogData = new DialogData('Confirm Action', 'Are you sure you want to remove this player from this team');
+        const dialogRef = this.dialog.open(DialogCustomComponent, { maxWidth: '300px', data: dialogData });
 
-        const index = this.dataSource.data.indexOf(rowElement);
-        this.dataSource.data.splice(index, 1);
-        this.dataSource._updateChangeSubscription();
+        dialogRef.afterClosed().subscribe(dialogResult => {
+            if (dialogResult) {
+                this.teamService.deleteUserFromTeam(rowElement.uid);
+                const index = this.dataSource.data.indexOf(rowElement);
+                this.dataSource.data.splice(index, 1);
+                this.dataSource._updateChangeSubscription();
+            }
+        });
+
+
     }
 }
